+++ src/Receiver.cpp
@@ -114,15 +114,13 @@ Receiver::Receiver(HardwareSerial &uart)
       inactiveBuf_(buf2_),
       activeBufIndex_(0),
       packetSize_(0),
-      packetTimestamp_(0),
+      packetStats_{},
       lastBreakStartTime_(0),
       breakStartTime_(0),
       lastSlotEndTime_(0),
       connected_(false),
       connectChangeFunc_{nullptr},
-      packetTimeoutCount_(0),
-      framingErrorCount_(0),
-      shortPacketCount_(0),
+      errorStats_{},
       responderCount_(0),
       responderOutBufLen_{0},
       setTXNotRXFunc_(nullptr) {
@@ -377,7 +375,11 @@ void Receiver::begin() {
     return;
   }
 
+  // Reset all the stats
   resetPacketCount();
+  packetSize_ = 0;
+  packetStats_ = PacketStats{};
+  errorStats_ = ErrorStats{};
 
   // Set up the instance for the ISRs
   Receiver *r = rxInstances[serialIndex_];
@@ -598,14 +600,11 @@ void Receiver::end() {
   setConnected(false);
 }
 
-int Receiver::readPacket(uint8_t *buf, int startChannel, int len) {
-  if (packetSize_ <= 0) {
-    return -1;
-  }
+int Receiver::readPacket(uint8_t *buf, int startChannel, int len,
+                         PacketStats *stats) {
   if (len <= 0 || startChannel < 0 || kMaxDMXPacketSize <= startChannel) {
     return 0;
   }
-
 
   int retval = -1;
   Lock lock{*this};
@@ -625,14 +624,14 @@ int Receiver::readPacket(uint8_t *buf, int startChannel, int len) {
       }
       packetSize_ = 0;
     }
+    if (stats != nullptr) {
+      *stats = packetStats_;
+    }
   //}
   return retval;
 }
 
 uint8_t Receiver::get(int channel) const {
-  if (lastPacketSize_ <= 0) {
-    return 0;
-  }
   if (channel < 0 || kMaxDMXPacketSize <= channel) {
     return 0;
   }
@@ -641,7 +640,7 @@ uint8_t Receiver::get(int channel) const {
   Lock lock{*this};
   //{
     // Since channel >= 0, lastPacketSize_ > channel implies lastPacketSize_ > 0
-    if (channel < lastPacketSize_) {
+    if (channel < packetStats_.size) {
       b = inactiveBuf_[channel];
     }
   //}
@@ -652,9 +651,6 @@ uint16_t Receiver::get16Bit(int channel, bool *rangeError) const {
   if (rangeError != nullptr) {
     *rangeError = true;
   }
-  if (lastPacketSize_ <= 0) {
-    return 0;
-  }
   if (channel < 0 || kMaxDMXPacketSize - 1 <= channel) {
     return 0;
   }
@@ -664,7 +660,7 @@ uint16_t Receiver::get16Bit(int channel, bool *rangeError) const {
   //{
     // Since channel >= 0, lastPacketSize_ - 1 > channel
     // implies lastPacketSize_ - 1 > 0
-    if (channel < lastPacketSize_ - 1) {
+    if (channel < packetStats_.size - 1) {
       if (rangeError != nullptr) {
         *rangeError = false;
       }
@@ -760,8 +756,13 @@ void Receiver::completePacket() {
   }
 
   incPacketCount();
-  lastPacketSize_ = packetSize_ = activeBufIndex_;
-  packetTimestamp_ = t;
+  packetStats_.size = packetSize_ = activeBufIndex_;
+  packetStats_.timestamp = t;
+  if (lastSlotEndTime_ >= breakStartTime_) {
+    packetStats_.packetTime = lastSlotEndTime_ - breakStartTime_;
+  } else {
+    packetStats_.packetTime = 0;
+  }
 
   // Let the responder, if any, process the packet
   if (responders_ != nullptr) {
@@ -769,7 +770,7 @@ void Receiver::completePacket() {
     if (r != nullptr) {
       r->receivePacket(inactiveBuf_, packetSize_);
       if (r->eatPacket()) {
-        lastPacketSize_ = packetSize_ = 0;
+        packetStats_.size = packetSize_ = 0;
       }
     }
   }
@@ -783,14 +784,14 @@ void Receiver::checkPacketTimeout() {
   if (state_ == RecvStates::kBreak) {
     // This catches the case where a short BREAK is followed by a longer MAB
     if ((t - breakStartTime_) < 88 + 44) {
-      framingErrorCount_++;
+      errorStats_.framingErrorCount++;
       completePacket();
       setConnected(false);
     }
   } else if (state_ == RecvStates::kData) {
     if ((t - breakStartTime_) > kMaxDMXPacketTime ||
         (t - lastSlotEndTime_) >= kMaxDMXIdleTime) {
-      packetTimeoutCount_++;
+      errorStats_.packetTimeoutCount++;
       completePacket();
       setConnected(false);
     }
@@ -815,7 +816,7 @@ void Receiver::receivePotentialBreak() {
 
 void Receiver::receiveBadBreak() {
   // Not a break
-  framingErrorCount_++;
+  errorStats_.framingErrorCount++;
 
   // Don't keep the packet
   // See: [BREAK timing at the receiver](http://www.rdmprotocol.org/forums/showthread.php?t=1292)
@@ -849,19 +850,21 @@ void Receiver::receiveByte(uint8_t b) {
                                 // seen some timeout
         // Complete any un-flushed bytes
         uint32_t dt = breakStartTime_ - lastBreakStartTime_;
+        packetStats_.breakToBreakTime = dt;
         if (dt < kMinDMXPacketTime) {
-          shortPacketCount_++;
+          errorStats_.shortPacketCount++;
           // Discard the data
           activeBufIndex_ = 0;
         } else if (dt > kMaxDMXPacketTime) {
           // NOTE: Zero-length packets will also trigger a timeout
-          packetTimeoutCount_++;
+          errorStats_.packetTimeoutCount++;
           // Keep the data
         }
         completePacket();
       } else {
         activeBufIndex_ = 0;
       }
+      packetStats_.breakPlusMABTime = eopTime - 44 - breakStartTime_;
       lastBreakStartTime_ = breakStartTime_;
       setConnected(true);
       state_ = RecvStates::kData;
@@ -892,7 +895,7 @@ void Receiver::receiveByte(uint8_t b) {
   // until, but not including, this one
   lastSlotEndTime_ = eopTime;
   if ((eopTime - breakStartTime_) > kMaxDMXPacketTime) {
-    packetTimeoutCount_++;
+    errorStats_.packetTimeoutCount++;
     completePacket();
     return;
   }